requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
convertdate==2.2.2
cssselect==1.1.0
cssutils==1.0.2
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
holidays==0.10.3
idna==2.10
korean-lunar-calendar==0.2.1
lxml==4.6.0
numpy==1.19.2
pandas==1.1.3
premailer==3.7.0
psycopg2==2.8.6
PyMeeus==0.3.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.10
yagmail==0.14.245
yahoofinancials==1.5
django-background-tasks=1.2.5